Value	Store time as floating point. 	Missed conversions.
svn path=/trunk/koffice/; revision=617764

+++ kspread/ValueConverter.cpp
@@ -320,7 +320,7 @@ Value ValueConverter::asTime (const Value &value) const
       val.setValue ( Value( QTime::currentTime(), doc() ) );
     break;
     case Value::Integer:
-      val.setValue (value.asInteger());
+      val.setValue (value.asFloat());
       val.setFormat (Value::fmt_Time);
     break;
     case Value::Float:
